Refactor RepositoryVersion context manager, so it lives in pulpcore. Bring back the RepositoryVersion create endpoint.
re #3285
https://pulp.plan.io/issues/3285
