+++ pulpcore/pulpcore/app/wrappers/__init__.py
@@ -1,0 +1,1 @@
+from .repository import RepositoryVersion  # noqa
